requirements-dev.txt
-i https://pypi.org/simple

astroid==2.15.6 ; python_full_version >= '3.7.2'

certifi==2023.7.22 ; python_version >= '3.6'

charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'

colorama==0.4.6 ; sys_platform == 'win32'

dill==0.3.6 ; python_version < '3.11'

docutils==0.20.1 ; python_version >= '3.7'

flit==3.9.0

flit-core==3.9.0 ; python_version >= '3.6'

idna==3.4 ; python_version >= '3.5'

iniconfig==2.0.0 ; python_version >= '3.7'

isort==5.12.0

lazy-object-proxy==1.9.0 ; python_version >= '3.7'

mccabe==0.7.0 ; python_version >= '3.6'

mypy==1.5.1

mypy-extensions==1.0.0 ; python_version >= '3.5'

packaging==23.1 ; python_version >= '3.7'

platformdirs==3.10.0 ; python_version >= '3.7'

pluggy==1.3.0 ; python_version >= '3.6'

pylint==2.17.5

pytest==7.4.0

pytest-asyncio==0.21.1

requests==2.31.0

toml==0.10.2

tomli==2.0.1 ; python_version < '3.11'

tomli-w==1.0.0 ; python_version >= '3.7'

tomlkit==0.12.1 ; python_version >= '3.7'

typing-extensions==4.7.1 ; python_version >= '3.7'

urllib3==2.0.4 ; python_version >= '3.7'

wrapt==1.15.0 ; python_version < '3.11'

yapf==0.40.1

std-uritemplate==0.0.35

opentelemetry-api==1.19.0

opentelemetry-sdk==1.19.0

